Conflicts:
	ProjectSDL2.1/GLModel.cpp
	ProjectSDL2.1/ProjectSDL2.1.vcxproj
	ProjectSDL2.1/obj_loader.cpp

+++ ProjectSDL2.1/GLModel.cpp
@@ -38,7 +38,7 @@ GLModel::GLModel(FishBox* FSH_Loader, unsigned int modelID)
 
 GLModel::~GLModel()
 {
-	for (int i = 0; i < meshes.size(); i++)
+	for (size_t i = 0; i < meshes.size(); i++)
 	{
 		delete meshes[i];
 	}
@@ -48,7 +48,7 @@ GLModel::~GLModel()
 void GLModel::Draw(GLShader& shader)
 {
 
-	for (int i = 0; i < meshes.size(); i++)
+	for (size_t i = 0; i < meshes.size(); i++)
 	{
 		glUniform3fv(shader.GetUnifromLocation("diffuse"), 1, glm::value_ptr(glm::vec3(meshes[i]->GetMaterial()->diffuse[0], meshes[i]->GetMaterial()->diffuse[1], meshes[i]->GetMaterial()->diffuse[2])));
 		glUniform3fv(shader.GetUnifromLocation("ambient"), 1, glm::value_ptr(glm::vec3(meshes[i]->GetMaterial()->ambient[0], meshes[i]->GetMaterial()->ambient[1], meshes[i]->GetMaterial()->ambient[2])));
